Update patch set 1
Patch Set 1:

Fixed the whitespace (cut-and-paste job originally). Added a bit more context to the intro so (hopefully) it's clearer what CRAMS is and how this bit of work fits in.

Also, worth noting that this particular piece of work already had specs drawn up and agreed outside of the nectar-specs repo/process, but I thought we should put this in here as a place for further discussions on it and to make it public to rc-ops etc.

Patch-set: 1
Label: Verified=0