Update patch set 1
Patch Set 1:

> Uploaded patch set 1.

I'm not expecting us to merge this as it stands - I made manual modifications for this mockup to work on local storage, which aren't appropriate for the real Dashboard. I just thought it would be convenient to put this into Gerrit also.

The widget as it stands is almost as minimal as possible. The header can be removed, but then we have to carefully adhere to the Twitter display guidelines - most especially, we'd have to add a Twitter logo somewhere else, and we'd have to clearly attribute the content to Twitter. If we leave the header as it looks here, we automatically comply with the guidelines, AFAICT.

Patch-set: 1
Reviewer: Gerrit User 1000139 <1000139@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Verified=0